Add navigation buttons and animationContainer components
